# server/openapi.yaml
openapi: 3.0.3
info:
  title: Apps auth
  version: 0.1.0
  description: Shared sign-in for the personal app site.
paths:
...
  /api/session:
    get:
      operationId: getSession
      responses:
        "200":
          description: The current session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "401":
          $ref: "#/components/responses/Error"
...
  /api/users:
    get:
      operationId: listUsers
      responses:
        "200":
          description: Accounts an admin can manage.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserList"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
    post:
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: Created. New accounts are regular users with no apps.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/users/{id}:
    patch:
      operationId: updateUser
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /api/users/{id}/apps:
    put:
      operationId: setUserApps
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppListRequest"
      responses:
        "200":
          description: Replaced app list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Error:
      description: Request failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    Session:
      type: object
      required: [kind, apps, expiresAt]
      properties:
        kind:
          type: string
          enum: [user, temporary]
        username:
          type: string
        nickname:
          type: string
        role:
          type: string
          enum: [admin, user]
        apps:
          type: array
          items:
            type: string
        expiresAt:
          type: string
          format: date-time
    User:
      type: object
      required: [id, username, role, disabled, apps]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        role:
          type: string
          enum: [admin, user]
        disabled:
          type: boolean
        apps:
          type: array
          items:
            type: string
    UserList:
      type: object
      required: [users]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
    CreateUserRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    UpdateUserRequest:
      type: object
      properties:
        password:
          type: string
        disabled:
          type: boolean
        role:
          type: string
          enum: [admin, user]
    AppListRequest:
      type: object
      required: [slugs]
      properties:
        slugs:
          type: array
          items:
            type: string
...